refactor(view): 입력받을 때 사용하는 상수와 게임 진행에 필요한 상수 분리
view에서만 사용하는 상수를 view 패키지 내로 분리시켰습니다.
InputParser의 패키지를 view로 이동했습니다.
