+++ src/main/java/view/constant/ViewRule.java
@@ -1,0 +1,10 @@
+package view.constant;
+
+import java.util.List;
+
+public class ViewRule {
+    public static final String LINE_SEPARATOR = "\n";
+
+    public static final List<String> YES_ANSWER = List.of("Y", "y");
+    public static final List<String> NO_ANSWER = List.of("N", "n");
+}
